Fixed mentioned remarks
